New arm gesture with wirst movement
With the current arm gesture you have to swtich off the saber to store it vertically (e.g. in the saber bag).
WIth this new gesture you have to hold the saber vertical AND perform a wirst movement (i.e. accelerate on a motorbike).
The threshold could be fine tuned if necessary.